re-integration event class
